+++ itsik/code/v26/accuracy_funcs.py
@@ -90,7 +90,6 @@ def multi_label_accuracy_weighted_loss(outs, samples, nclasses):
     """    
     preds, task_accuracy = multi_label_accuracy_base(outs, samples, nclasses)
     loss_weight = samples.loss_weight
-    print(f'{task_accuracy.shape} {loss_weight.shape}')
     task_accuracy = task_accuracy * loss_weight
     # It was in the original version only regular mean in the second access
     # we want to do a weighted mean
