+++ itsik/code/v26/models/flag_at.py
@@ -1,0 +1,10 @@
+from enum import Enum, auto
+
+
+class FlagAt(Enum):
+    BU1 = auto()
+    TD = auto()
+    BU2 = auto()
+    NOFLAG = auto()
+    BU1_SIMPLE = auto()
+    BU1_NOLAT = auto()
